+++ sdk/python/agentfileops/__init__.py
@@ -1,0 +1,3 @@
+from .client import AgentFileOpsError, Client
+
+__all__ = ["AgentFileOpsError", "Client"]